Move interface definition for `IPositronWebviewPreloadService` to browser/
