+++ src/vs/workbench/contrib/positronWebviewPreloads/browser/utils.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { ILanguageRuntimeMessageOutput } from 'vs/workbench/services/languageRuntime/common/languageRuntimeService';
-import { MIME_TYPE_HOLOVIEWS_LOAD, MIME_TYPE_HOLOVIEWS_EXEC, MIME_TYPE_BOKEH_EXEC, MIME_TYPE_BOKEH_LOAD, MIME_TYPE_POSITRON_WEBVIEW_FLAG } from 'vs/workbench/services/positronWebviewPreloads/common/positronWebviewPreloadService';
+import { MIME_TYPE_HOLOVIEWS_LOAD, MIME_TYPE_HOLOVIEWS_EXEC, MIME_TYPE_BOKEH_EXEC, MIME_TYPE_BOKEH_LOAD, MIME_TYPE_POSITRON_WEBVIEW_FLAG } from 'vs/workbench/services/positronWebviewPreloads/browser/positronWebviewPreloadService';
 
 
 const webviewReplayMimeTypes = new Set([
